Demote KernelInventory to TwoLevelStrategy's private splice watcher
It was presented as a composable component (its own module, a peer of the
discovered strategies) but its only consumer is the two-level strategy's
enrollment. Folded into two_level.py as _KernelInventory; it still derives
from PipelineStrategy because Run.strategies is the channel the engine
notifies — the event protocol is how a search shape hears about splices —
but it is no longer public surface.
